set license
