+++ LICENSE
@@ -1,0 +1,3 @@
+This project incorporates code from the [Gno repository](https://github.com/gnolang/gno), which is licensed under the terms outlined in its [LICENSE](https://github.com/gnolang/gno/blob/master/LICENSE.md).
+
+The use of the Gno code in this repository complies with the licensing terms defined in the original LICENSE file. You can find the full license text at the link above.
